Initial Setup for the dev branch
To start off I have added a game scene and a menu scene. The game scene has a tile map with a character that can move around, there is also a pause menu. The menu scene has a menu where you can start the game and go to option menu.

Inside the dev branch we can create version that is stable before pushing it to  the main branch.